+++ deno/payloads/v10/mod.ts
@@ -2,6 +2,7 @@ export * from '../common.ts';
 export * from './achievement.ts';
 export * from './application.ts';
 export * from './auditLog.ts';
+export * from './autoModeration.ts';
 export * from './channel.ts';
 export * from './emoji.ts';
 export * from './gateway.ts';