+++ payloads/v10/index.ts
@@ -2,6 +2,7 @@ export * from '../common';
 export * from './achievement';
 export * from './application';
 export * from './auditLog';
+export * from './autoModeration';
 export * from './channel';
 export * from './emoji';
 export * from './gateway';