+++ rest/v10/guild.ts
@@ -53,6 +53,9 @@ export type APIGuildCreatePartialChannel = StrictPartial<
 		| 'rtc_region'
 		| 'video_quality_mode'
 		| 'flags'
+		| 'default_reaction_emoji'
+		| 'available_tags'
+		| 'default_sort_order'
 	>
 > &
 	AddUndefinedToPossiblyUndefinedPropertiesOfInterface<{
@@ -135,7 +138,7 @@ export type RESTPostAPIGuildsJSONBody = AddUndefinedToPossiblyUndefinedPropertie
 	 */
 	afk_channel_id?: number | Snowflake | null;
 	/**
-	 * AFK timeout in seconds
+	 * afk timeout in seconds, can be set to: `60`, `300`, `900`, `1800`, `3600`
 	 */
 	afk_timeout?: number;
 	/**
@@ -235,7 +238,7 @@ export type RESTPatchAPIGuildJSONBody = AddUndefinedToPossiblyUndefinedPropertie
 	 */
 	afk_channel_id?: Snowflake | null;
 	/**
-	 * AFK timeout in seconds
+	 * afk timeout in seconds, can be set to: `60`, `300`, `900`, `1800`, `3600`
 	 */
 	afk_timeout?: number;
 	/**
@@ -357,7 +360,7 @@ export type RESTPatchAPIGuildChannelPositionsJSONBody = Array<
 export type RESTPatchAPIGuildChannelPositionsResult = never;
 
 /**
- * https://discord.com/developers/docs/resources/guild#list-active-threads
+ * https://discord.com/developers/docs/resources/guild#list-active-guild-threads
  */
 export type RESTGetAPIGuildThreadsResult = APIThreadList;
 
@@ -573,8 +576,14 @@ export type RESTGetAPIGuildBanResult = APIBan;
 export type RESTPutAPIGuildBanJSONBody = AddUndefinedToPossiblyUndefinedPropertiesOfInterface<{
 	/**
 	 * Number of days to delete messages for (0-7)
+	 *
+	 * @deprecated use `delete_message_seconds` instead
 	 */
 	delete_message_days?: number;
+	/**
+	 * Number of seconds to delete messages for, between 0 and 604800 (7 days)
+	 */
+	delete_message_seconds?: number;
 }>;
 
 /**
@@ -855,7 +864,7 @@ export type RESTPatchAPIGuildVoiceStateCurrentMemberJSONBody = AddUndefinedToPos
 	/**
 	 * The id of the channel the user is currently in
 	 */
-	channel_id: Snowflake;
+	channel_id?: Snowflake;
 	/**
 	 * Toggles the user's suppress state
 	 */